backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
boto3==1.35.63
botocore==1.35.99
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
dj-rest-auth==7.0.2
Django==4.2.27
django-allauth==65.13.1
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
